johan_defconfig: add custom config
Add a custom trimmed configuration that can be used to test SC8280XP and
X1E80100 support including the Lenovo ThinkPad X13s, sc8280xp-crd,
sa8295p-adp and x1e80100-crd.

Make sure the initramfs includes any modules required to boot, for
example:

	nvme phy_qcom_qmp_pcie pcie_qcom

for the X13s,

	nvme tcsrcc_x1e80100 phy_qcom_qmp_pcie pcie_qcom

for the x1e80100-crd, and

	phy_qcom_qmp_ufs ufs_qcom

for the sc8280xp-crd with rootfs on UFS.

For keyboard input and (more than 30 seconds of) display in initramfs,
make sure to also include:

	i2c_hid_of i2c_qcom_geni

for keyboard, and

	leds_qcom_lpg pwm_bl
	qrtr pmic_glink_altmode gpio_sbu_mux phy_qcom_qmp_combo
	gpucc_sc8280xp dispcc_sc8280xp
	phy_qcom_edp panel_edp msm

for the display.

For more details see:

	https://github.com/jhovold/linux/wiki/X13s

Not-signed-off-by: Johan Hovold <johan+linaro@kernel.org>