johan_defconfig: add custom config
Add a custom trimmed configuration that can be used to test SC8280XP
support (e.g. ADP, CRD and X13s).

Note that the default CMA size is increased to 128 MB due to NVMe
(see CONFIG_CMA_SIZE_MBYTES).

Make sure the initramfs includes any modules required to boot, for
example:

	phy_qcom_qmp_ufs ufs_qcom

for the CRD and

	nvme phy_qcom_qmp_pcie

for the X13s.

For keyboard input and (more than 30 seconds of) display in initramfs,
make sure to also include:

	i2c_hid_of i2c_qcom_geni

for keyboard, and

	leds_qcom_lpg pwm_bl
	phy_qcom_qmp_combo

for the display.

Note that the kernel command line needs to include the following
arguments for now:

	clk_ignore_unused pd_ignore_unused arm64.nopauth cma=128M

Depending on the firmware revision you may also need

	efi=noruntime

to be able to power off and reboot.

Not-signed-off-by: Johan Hovold <johan+linaro@kernel.org>