johan_defconfig: add custom config
Add a custom trimmed configuration that can be used to test SC8280XP
support (e.g. ADP, CRD and X13s).

Note that the default CMA size is increased to 128 MB due to NVMe
(CONFIG_CMA_SIZE_MBYTES).

Make sure the initramfs includes any modules required to boot, for
example:

	phy_qcom_qmp_ufs ufs_qcom

for the CRD and

	nvme phy_qcom_qmp_pcie pcie_qcom

for the X13s.

For a functional rescue shell in initramfs, make sure to also include

	i2c_hid_of i2c_qcom_geni
	leds_qcom_lpg pwm_bl

for keyboard and backlight, and

	qrtr pmic_glink_altmode gpio_sbu_mux phy_qcom_qmp_combo
	panel-edp msm phy_qcom_edp

for the display (preferably in the above order to avoid an msm drm
probe-deferral issue).

Not-signed-off-by: Johan Hovold <johan+linaro@kernel.org>