hack: interconnect: qcom: sc8280xp: Disable syncstate for now
While the ADP boots reliably (with just the PCIe keepalive hack), the
sync_state shuts down the busses needed for the efifb to display the
console on the CRD - with system instability resulting.

Disable sync_state until we have adequate clients voting for resources
to keep things alive and stable.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>