hack: ath11k: override qmi board id
Force the ath11k driver to load a board file for a different machine
that is already in linux-firmware.

For proper WiFi support we need a board file from the vendor for the
actual hardware in question.

WARNING: This should only be used for basic internal testing. Use at
your own risk.

Not-signed-off-by: Johan Hovold <johan+linaro@kernel.org>