hack: firmware: efi: disable reset service
The EFI reset services crashes on reboot and power off. Disable it so
that we fall back to PSCI.

This is needed until the boot firmware provides an RT_PROP table that
describes the supported services. An alternative is to disable all
runtime services by booting with 'efi=noruntime', but that would also
prevent access to the UEFI variables.

Not-signed-off-by: Johan Hovold <johan+linaro@kernel.org>