arm64: dts: qcom: sc8280xp-x13s: model the PMU of the on-board wcn6855
Add a node for the PMU of the WCN6855 and rework the inputs of the wifi
and bluetooth nodes to consume the PMU's outputs.

With this we can drop the regulator-always-on properties from vreg_s11b
and vreg_s12b as they will now be enabled by the power sequencing
driver.

Tested-by: Steev Klimaszewski <steev@kali.org> # Thinkpad X13s
Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>
Link: https://lore.kernel.org/r/20240930103041.49229-4-brgl@bgdev.pl
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>