WIP: remoteproc: q6v5-pas: Attach to lite ADSP firmware
If we suspect to have a "lite" version of the firmware already loaded by
the earlier boot firmware, then keep it running and report this as
RPROC_DETACHED state to the remoteproc subsystem. That way, we probe glink
channels, services etc as usual and can use them until the remoteproc is
later restarted with the full firmware.

TODO: Where and who should trigger the restart necessary to load the full
firmware?

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>