arm64: dts: qcom: x1-el2: Add iris video-firmware node
Add the video-firmware node with the IOMMU used for loading the Iris
firmware to make video acceleration work when running in EL2.

In EL2 we can load the generic firmware, so there is no need to keep Iris
disabled by default. The generic firmware is used to validate the Iris
driver and will often be more reliable, so override the firmware-name if
needed to use that even if a device-specific firmware-name is defined.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>