wip: media: iris: Port firmware loading without TZ/PAS from venus
Add support to load firmware on devices running bare-metal without support
for the firmware PAS service in the TZ or hypervisor firmware. The code is
largely copied as-is from venus/firmware.c, except with the old A9SS
sequence removed.

The "non_tz" boot sequence is used when there is a "video-firmware" subnode
defined in the device tree with the IOMMU for the Iris firmware.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>